# src/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CryptoTrace REST API",
    "description": "Cryptographic Fingerprinting & Data Classification Engine — REST API",
    "version": "0.1.0",
    "contact": {
      "name": "CryptoTrace Team",
      "url": "https://github.com/cryptotrace/cryptotrace"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "servers": [
    { "url": "http://localhost:8080", "description": "Local development" }
  ],
  "paths": {
...
    "/v1/jobs/{id}": {
      "get": {
        "summary": "Get job status and result",
        "description": "Poll a submitted job. Returns 200 with result when complete.",
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Job info (may include result if completed)"
          },
          "404": { "description": "Job not found" }
        }
      },
      "delete": {
        "summary": "Cancel or remove a job",
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Job cancelled or removed" },
          "404": { "description": "Job not found" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for authentication. Can also be provided as Authorization: Bearer <key>."
      }
    },
...
  }
}